backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
redis==5.1.1
celery==5.4.0
boto3==1.35.29
structlog==24.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
opencv-python-headless==4.10.0.84
numpy==2.1.1
ultralytics==8.3.0
ffmpeg-python==0.2.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
